refactor: promote mtgo_bridge to services/mtgo_bridge_service
The MTGO CLI bridge is a stateful external-process integration with a
subprocess/multiprocessing transport plus a Python facade -- structurally
identical to BundleSnapshotClient/RemoteSnapshotClient, which are services.
utils is documented as cross-cutting helpers only, so this fits the
services layer better.

- utils/mtgo_bridge_client.py -> services/mtgo_bridge_service/client.py
  (transport)
- utils/mtgo_bridge.py -> services/mtgo_bridge_service/__init__.py (facade)
- update scripts, collection_service, timer_alert, and the bridge test
- refresh ARCHITECTURE.md / README.md / diagram generator (also reflects
  the deck_text_cache -> repositories and results_filter moves)

Tradeoff: test_mtgo_bridge_client.py is no longer headless-importable
since it now transitively imports the services package (which already
hard-requires wx via bundle_snapshot_client). Consistent with the rest
of the services test suite; tests run on the wx-enabled Windows side.
